add matrix and fix monoid

+++ group/monoid/add-count.hpp
@@ -8,13 +8,13 @@
 
 template<class T>
 struct AddCountMonoid{
-    using P = pair<T,int>;
+    using P = pair<T,ll>;
     using value_type = P;
     static constexpr P op(const P &x,const P &y){
         return P(x.first+y.first,x.second+y.second);
     }
     static constexpr P inverse(const P &x){return P(-x.first,-x.second);}
-    static constexpr P unit(){return P(T(0),0);}
-    static constexpr P make(const T &x){return P(x,1);}
+    static constexpr P unit(){return P(T(0),0LL);}
+    static constexpr P make(const T &x){return P(x,1LL);}
 };
